+++ src/main/java/lotto/exception/LottoGameException.java
@@ -1,4 +1,4 @@
-package lotto.util;
+package lotto.exception;
 
 import lotto.constance.GameConst;
