+++ src/main/java/lotto/exception/exceptionhandler/ExceptionHandler.java
@@ -1,4 +1,4 @@
-package lotto.exceptionhandler;
+package lotto.exception.exceptionhandler;
 
 import java.util.function.Supplier;
