+++ src/main/java/lotto/model/domain/Results.java
@@ -2,7 +2,6 @@ package lotto.model.domain;
 
 import java.util.Arrays;
 import java.util.Collections;
-import java.util.Comparator;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
