+++ src/test/java/lotto/exceptionhandler/RetryExceptionHandlerTest.java
@@ -1,9 +1,9 @@
 package lotto.exceptionhandler;
 
-import static org.junit.jupiter.api.Assertions.*;
-
 import java.util.LinkedList;
 import java.util.List;
+import lotto.exception.exceptionhandler.ExceptionHandler;
+import lotto.exception.exceptionhandler.RetryExceptionHandler;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
