+++ src/test/java/lotto/model/domain/LottoAnswerTest.java
@@ -1,8 +1,7 @@
 package lotto.model.domain;
 
-import static org.junit.jupiter.api.Assertions.*;
-
 import java.util.List;
+import lotto.model.domain.lotto.LottoAnswer;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
