+++ src/test/java/lotto/util/LottoGameExceptionTest.java
@@ -1,9 +1,7 @@
 package lotto.util;
 
-import static org.junit.jupiter.api.Assertions.*;
-
+import lotto.exception.LottoGameException;
 import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
